requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-oauth-toolkit==2.2.0
drf-spectacular==0.26.4
